Move get_aggregated_result out of checkengine
Error handling / impedance mismatch -> should happen at the boundary.

CMK-12298
