+++ cmk/checkengine/checking/__init__.py
@@ -4,7 +4,14 @@
 # conditions defined in the file COPYING, which is part of this source code package.
 
 from ._checking import CheckPluginName, CheckPluginNameStr, Item
-from ._plugin import CheckPlugin
+from ._plugin import AggregatedResult, CheckPlugin
 from ._timing import make_timing_results
 
-__all__ = ["CheckPlugin", "CheckPluginName", "CheckPluginNameStr", "Item", "make_timing_results"]
+__all__ = [
+    "AggregatedResult",
+    "CheckPlugin",
+    "CheckPluginName",
+    "CheckPluginNameStr",
+    "Item",
+    "make_timing_results",
+]
